feat: first draft
