+++ package.json
@@ -1,63 +1,63 @@
 {
-  "name": "package_name",
-  "version": "0.0.0",
-  "description": "package_description",
-  "keywords": [
-    "typescript",
-    "prismic"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "ssh://git@github.com/github_org_slash_github_repo.git"
-  },
-  "license": "Apache-2.0",
-  "author": "Prismic <contact@prismic.io> (https://prismic.io)",
-  "type": "module",
-  "exports": {
-    ".": {
-      "require": "./dist/index.cjs",
-      "import": "./dist/index.mjs"
-    },
-    "./package.json": "./package.json"
-  },
-  "main": "dist/index.cjs",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "src"
-  ],
-  "scripts": {
-    "build": "siroc build",
-    "dev": "siroc build --watch",
-    "format": "prettier --write .",
-    "prepare": "npm run build",
-    "release": "yarn build && yarn test && standard-version && git push --follow-tags && yarn build && npm publish",
-    "release:dry": "standard-version --dry-run",
-    "release:alpha": "yarn build && yarn test && standard-version --release-as major --prerelease alpha && git push --follow-tags && yarn build && npm publish --tag alpha",
-    "release:alpha:dry": "standard-version --release-as major --prerelease alpha --dry-run",
-    "lint": "eslint --ext .js,.ts .",
-    "unit": "nyc --reporter=lcovonly --reporter=text --exclude-after-remap=false ava",
-    "test": "yarn lint && yarn unit"
-  },
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.26.1",
-    "@typescript-eslint/parser": "^4.26.1",
-    "ava": "^3.15.0",
-    "eslint": "^7.28.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "nyc": "^15.1.0",
-    "prettier": "^2.3.1",
-    "siroc": "^0.11.0",
-    "standard-version": "^9.3.0",
-    "ts-eager": "^2.0.2",
-    "typescript": "^4.3.2"
-  },
-  "engines": {
-    "node": ">=12.7.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
+	"name": "@prismicio/richtext",
+	"version": "0.0.0",
+	"description": "Convert Rich Text raw data to a generic tree for serialization",
+	"keywords": [
+		"typescript",
+		"prismic"
+	],
+	"repository": {
+		"type": "git",
+		"url": "ssh://git@github.com/prismicio/prismic-richtext.git"
+	},
+	"license": "Apache-2.0",
+	"author": "Prismic <contact@prismic.io> (https://prismic.io)",
+	"type": "module",
+	"exports": {
+		".": {
+			"require": "./dist/index.cjs",
+			"import": "./dist/index.mjs"
+		},
+		"./package.json": "./package.json"
+	},
+	"main": "dist/index.cjs",
+	"module": "dist/index.mjs",
+	"types": "dist/index.d.ts",
+	"files": [
+		"dist",
+		"src"
+	],
+	"scripts": {
+		"build": "siroc build",
+		"dev": "siroc build --watch",
+		"format": "prettier --write .",
+		"prepare": "npm run build",
+		"release": "npm run build && npm run test && standard-version && git push --follow-tags && npm run build && npm publish",
+		"release:dry": "standard-version --dry-run",
+		"release:alpha": "npm run build && npm run test && standard-version --release-as major --prerelease alpha && git push --follow-tags && npm run build && npm publish --tag alpha",
+		"release:alpha:dry": "standard-version --release-as major --prerelease alpha --dry-run",
+		"lint": "eslint --ext .js,.ts .",
+		"unit": "nyc --reporter=lcovonly --reporter=text --exclude-after-remap=false ava",
+		"test": "npm run lint && npm run unit"
+	},
+	"devDependencies": {
+		"@typescript-eslint/eslint-plugin": "^4.26.1",
+		"@typescript-eslint/parser": "^4.26.1",
+		"ava": "^3.15.0",
+		"eslint": "^7.28.0",
+		"eslint-config-prettier": "^8.3.0",
+		"eslint-plugin-prettier": "^3.4.0",
+		"nyc": "^15.1.0",
+		"prettier": "^2.3.1",
+		"siroc": "^0.11.0",
+		"standard-version": "^9.3.0",
+		"ts-eager": "^2.0.2",
+		"typescript": "^4.3.2"
+	},
+	"engines": {
+		"node": ">=12.7.0"
+	},
+	"publishConfig": {
+		"access": "public"
+	}
 }
