+++ src/lib/uuid.ts
@@ -6,4 +6,5 @@ interface UUIDFn {
 export const uuid: UUIDFn = (): string => {
 	return (++uuid.i).toString();
 };
+
 uuid.i = 0;
