+++ test/snapshots/asText.test.ts.md
@@ -1,0 +1,19 @@
+# Snapshot report for `test/asText.test.ts`
+
+The actual snapshot is saved in `asText.test.ts.snap`.
+
+Generated by [AVA](https://avajs.dev).
+
+## returns a string representation of a rich text field value
+
+> Snapshot 1
+
+    `This block has a line␊
+    break just before the word "break" Text before with a multi-block span! Heading 1 bold text italic still bold normal Paragraph A paragraph with a nested label First ordered list item Second ordered list item Third ordered list item Bold and now italic but not anymore value in the last list item Paragraph First list item Second list item Third list item Paragraph A web link A document link A media link A non-image media link A paragraph with a nested document link Paragraph Paragraph`
+
+## allows for custom separator
+
+> Snapshot 1
+
+    `This block has a line␊
+    break just before the word "break"SEPARATORText before with a multi-block span!SEPARATORHeading 1 bold text italic still bold normalSEPARATORParagraphSEPARATORA paragraph with a nested labelSEPARATORFirst ordered list itemSEPARATORSecond ordered list itemSEPARATORThird ordered list itemSEPARATORBold and now italic but not anymore value in the last list itemSEPARATORParagraphSEPARATORFirst list itemSEPARATORSecond list itemSEPARATORThird list itemSEPARATORParagraphSEPARATORA web linkSEPARATORA document linkSEPARATORA media linkSEPARATORA non-image media linkSEPARATORA paragraph with a nested document linkSEPARATORParagraphSEPARATORParagraph`
